refactor: moving usage-related modules
